Crowdsale: Added additional check to setEndsAt()
According to https://medium.com/@maraoz/f1d742b94756:
"Function setEndsAt in line 378 of Crowdsale.sol can leave the contract
in an invalid state, as this comment reads. This is generally not a good idea.
Consider adding post-conditions that check and throw if the contract is left
in an invalid state."

Now checking if "now" is earlier than endsAt, and therefore preventing
state confusions. Other possible state confusions should be checked, though.